pyproject.toml
[project]
name = "media-organizer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "tqdm>=4.66",
  "pillow>=10.0",
  "imagehash>=4.3",
  "exifread>=3.0",
]

dynamic = []

[project.scripts]
media-organizer = "media_organizer.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
